refactor: shared/api 레이어 구성
변경 사항:
- apis/ 폴더를 shared/api/로 이동
- axios 인스턴스 설정 추가 (shared/api/instance.ts)
- roomApi, userApi, answerApi를 shared/api/로 이동
- shared/api/index.ts에서 모든 API export

이유:
- FSD 아키텍처에 맞게 API 클라이언트를 shared 레이어로 통합
- axios 인스턴스를 통한 중앙 집중식 API 설정 관리
- 코드 재사용성 및 유지보수성 향상
