+++ client/src/shared/api/index.ts
@@ -1,3 +1,5 @@
+export * from './instance';
 export { default as roomApi } from './room.api';
 export { default as userApi } from './user.api';
 export { default as answerApi } from './answer.api';
+
